requirements.txt
# Use local for development only, not for deployment.
homeassistant~=2022.8.6
python-can~=4.0.0
voluptuous~=0.12
